Removed dots and inits.py's added a dht22 setup
